Solved Task 3

+++ app/controllers/products_controller.rb
@@ -15,9 +15,15 @@ class ProductsController < ApplicationController
   end
 
   def edit
+    if current_user != product.user
+      redirect_to category_product_url(category, product)
+      flash[:error] = 'You are not allowed to edit this product.'
+    end
   end
 
   def create
+    return redirect_to new_user_session_path unless user_signed_in?
+
     self.product = Product.new(product_params)
 
     if product.save
@@ -29,15 +35,26 @@ class ProductsController < ApplicationController
   end
 
   def update
-    if self.product.update(product_params)
-      redirect_to category_product_url(category, product), notice: 'Product was successfully updated.'
+    if user_signed_in?
+      if current_user == product.user
+        if self.product.update(product_params)
+          redirect_to category_product_url(category, product), notice: 'Product was successfully updated.'
+        else
+          render action: 'edit'
+        end
+      else
+        redirect_to category_product_url(category, product)
+        flash[:error] = "You are not allowed to edit this product."
+      end
     else
-      render action: 'edit'
+      redirect_to new_user_session_path
     end
   end
 
+
   # DELETE /products/1
   def destroy
+    return redirect_to new_user_session_path unless user_signed_in?
     product.destroy
     redirect_to category_url(product.category), notice: 'Product was successfully destroyed.'
   end
